+++ matrix-sdk/src/androidTest/java/org/matrix/androidsdk/crypto/keysbackup/KeysBackupTest.kt
@@ -417,6 +417,337 @@ class KeysBackupTest {
         testData.cryptoTestData.clear(context)
     }
 
+
+    /**
+     * - Do an e2e backup to the homeserver with a recovery key
+     * - And log Alice on a new device
+     * - The new device must see the previous backup as not trusted
+     * - Trust the backup from the new device
+     * - Backup must be enabled on the new device
+     * - Retrieve the last version from the server
+     * - It must be the same
+     * - It must be trusted and must have with 2 signatures now
+     */
+    @Test
+    fun trustKeyBackupVersionTest() {
+        // - Do an e2e backup to the homeserver with a recovery key
+        // - And log Alice on a new device
+        val context = InstrumentationRegistry.getContext()
+
+        val testData = createKeysBackupScenarioWithPassword(null, true)
+
+        val stateObserver = StateObserver(testData.aliceSession2.crypto!!.keysBackup)
+
+        // Wait for backup state to be NotTrusted
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.NotTrusted)
+
+        // - The new device must see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        // - Trust the backup from the new device
+        val latch = CountDownLatch(1)
+        testData.aliceSession2.crypto!!.keysBackup.trustKeysBackupVersion(
+                testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion!!,
+                true,
+                TestApiCallback(latch)
+        )
+        mTestHelper.await(latch)
+
+        // Wait for backup state to be ReadyToBackUp
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.ReadyToBackUp)
+
+        // - Backup must be enabled on the new device, on the same version
+        assertEquals(testData.prepareKeysBackupDataResult.version, testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion?.version)
+        assertTrue(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+
+        // - Retrieve the last version from the server
+        val latch2 = CountDownLatch(1)
+        var keysVersionResult: KeysVersionResult? = null
+        testData.aliceSession2.crypto!!.keysBackup.getCurrentVersion(
+                object : TestApiCallback<KeysVersionResult?>(latch2) {
+                    override fun onSuccess(info: KeysVersionResult?) {
+                        keysVersionResult = info
+                        super.onSuccess(info)
+                    }
+                }
+        )
+        mTestHelper.await(latch2)
+
+        // - It must be the same
+        assertEquals(testData.prepareKeysBackupDataResult.version, keysVersionResult!!.version)
+
+        val latch3 = CountDownLatch(1)
+        var keysBackupVersionTrust: KeysBackupVersionTrust? = null
+        testData.aliceSession2.crypto!!.keysBackup.getKeysBackupTrust(keysVersionResult!!,
+                object : TestApiCallback<KeysBackupVersionTrust>(latch3) {
+                    override fun onSuccess(info: KeysBackupVersionTrust) {
+                        keysBackupVersionTrust = info
+                        super.onSuccess(info)
+                    }
+                })
+        mTestHelper.await(latch3)
+
+        // - It must be trusted and must have 2 signatures now
+        assertTrue(keysBackupVersionTrust!!.usable)
+        assertEquals(2, keysBackupVersionTrust!!.signatures.size)
+
+        stateObserver.stopAndCheckStates(null)
+        testData.cryptoTestData.clear(context)
+    }
+
+    /**
+     * - Do an e2e backup to the homeserver with a recovery key
+     * - And log Alice on a new device
+     * - The new device must see the previous backup as not trusted
+     * - Trust the backup from the new device with the recovery key
+     * - Backup must be enabled on the new device
+     * - Retrieve the last version from the server
+     * - It must be the same
+     * - It must be trusted and must have with 2 signatures now
+     */
+    @Test
+    fun trustKeyBackupVersionWithRecoveryKeyTest() {
+        // - Do an e2e backup to the homeserver with a recovery key
+        // - And log Alice on a new device
+        val context = InstrumentationRegistry.getContext()
+
+        val testData = createKeysBackupScenarioWithPassword(null, true)
+
+        val stateObserver = StateObserver(testData.aliceSession2.crypto!!.keysBackup)
+
+        // Wait for backup state to be NotTrusted
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.NotTrusted)
+
+        // - The new device must see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        // - Trust the backup from the new device with the recovery key
+        val latch = CountDownLatch(1)
+        testData.aliceSession2.crypto!!.keysBackup.trustKeysBackupVersionWithRecoveryKey(
+                testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion!!,
+                testData.prepareKeysBackupDataResult.megolmBackupCreationInfo.recoveryKey,
+                TestApiCallback(latch)
+        )
+        mTestHelper.await(latch)
+
+        // Wait for backup state to be ReadyToBackUp
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.ReadyToBackUp)
+
+        // - Backup must be enabled on the new device, on the same version
+        assertEquals(testData.prepareKeysBackupDataResult.version, testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion?.version)
+        assertTrue(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+
+        // - Retrieve the last version from the server
+        val latch2 = CountDownLatch(1)
+        var keysVersionResult: KeysVersionResult? = null
+        testData.aliceSession2.crypto!!.keysBackup.getCurrentVersion(
+                object : TestApiCallback<KeysVersionResult?>(latch2) {
+                    override fun onSuccess(info: KeysVersionResult?) {
+                        keysVersionResult = info
+                        super.onSuccess(info)
+                    }
+                }
+        )
+        mTestHelper.await(latch2)
+
+        // - It must be the same
+        assertEquals(testData.prepareKeysBackupDataResult.version, keysVersionResult!!.version)
+
+        val latch3 = CountDownLatch(1)
+        var keysBackupVersionTrust: KeysBackupVersionTrust? = null
+        testData.aliceSession2.crypto!!.keysBackup.getKeysBackupTrust(keysVersionResult!!,
+                object : TestApiCallback<KeysBackupVersionTrust>(latch3) {
+                    override fun onSuccess(info: KeysBackupVersionTrust) {
+                        keysBackupVersionTrust = info
+                        super.onSuccess(info)
+                    }
+                })
+        mTestHelper.await(latch3)
+
+        // - It must be trusted and must have 2 signatures now
+        assertTrue(keysBackupVersionTrust!!.usable)
+        assertEquals(2, keysBackupVersionTrust!!.signatures.size)
+
+        stateObserver.stopAndCheckStates(null)
+        testData.cryptoTestData.clear(context)
+    }
+
+    /**
+     * - Do an e2e backup to the homeserver with a recovery key
+     * - And log Alice on a new device
+     * - The new device must see the previous backup as not trusted
+     * - Try to trust the backup from the new device with a wrong recovery key
+     * - It must fail
+     * - The backup must still be untrusted and disabled
+     */
+    @Test
+    fun trustKeyBackupVersionWithWrongRecoveryKeyTest() {
+        // - Do an e2e backup to the homeserver with a recovery key
+        // - And log Alice on a new device
+        val context = InstrumentationRegistry.getContext()
+
+        val testData = createKeysBackupScenarioWithPassword(null, true)
+
+        val stateObserver = StateObserver(testData.aliceSession2.crypto!!.keysBackup)
+
+        // Wait for backup state to be NotTrusted
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.NotTrusted)
+
+        // - The new device must see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        // - Try to trust the backup from the new device with a wrong recovery key
+        val latch = CountDownLatch(1)
+        testData.aliceSession2.crypto!!.keysBackup.trustKeysBackupVersionWithRecoveryKey(
+                testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion!!,
+                "Bad recovery key",
+                TestApiCallback(latch, false)
+        )
+        mTestHelper.await(latch)
+
+        // - The new device must still see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        stateObserver.stopAndCheckStates(null)
+        testData.cryptoTestData.clear(context)
+    }
+
+    /**
+     * - Do an e2e backup to the homeserver with a password
+     * - And log Alice on a new device
+     * - The new device must see the previous backup as not trusted
+     * - Trust the backup from the new device with the password
+     * - Backup must be enabled on the new device
+     * - Retrieve the last version from the server
+     * - It must be the same
+     * - It must be trusted and must have with 2 signatures now
+     */
+    @Test
+    fun trustKeyBackupVersionWithPasswordTest() {
+        val password = "Password"
+
+        // - Do an e2e backup to the homeserver with a password
+        // - And log Alice on a new device
+        val context = InstrumentationRegistry.getContext()
+
+        val testData = createKeysBackupScenarioWithPassword(password, true)
+
+        val stateObserver = StateObserver(testData.aliceSession2.crypto!!.keysBackup)
+
+        // Wait for backup state to be NotTrusted
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.NotTrusted)
+
+        // - The new device must see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        // - Trust the backup from the new device with the password
+        val latch = CountDownLatch(1)
+        testData.aliceSession2.crypto!!.keysBackup.trustKeysBackupVersionWithPassphrase(
+                testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion!!,
+                password,
+                TestApiCallback(latch)
+        )
+        mTestHelper.await(latch)
+
+        // Wait for backup state to be ReadyToBackUp
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.ReadyToBackUp)
+
+        // - Backup must be enabled on the new device, on the same version
+        assertEquals(testData.prepareKeysBackupDataResult.version, testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion?.version)
+        assertTrue(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+
+        // - Retrieve the last version from the server
+        val latch2 = CountDownLatch(1)
+        var keysVersionResult: KeysVersionResult? = null
+        testData.aliceSession2.crypto!!.keysBackup.getCurrentVersion(
+                object : TestApiCallback<KeysVersionResult?>(latch2) {
+                    override fun onSuccess(info: KeysVersionResult?) {
+                        keysVersionResult = info
+                        super.onSuccess(info)
+                    }
+                }
+        )
+        mTestHelper.await(latch2)
+
+        // - It must be the same
+        assertEquals(testData.prepareKeysBackupDataResult.version, keysVersionResult!!.version)
+
+        val latch3 = CountDownLatch(1)
+        var keysBackupVersionTrust: KeysBackupVersionTrust? = null
+        testData.aliceSession2.crypto!!.keysBackup.getKeysBackupTrust(keysVersionResult!!,
+                object : TestApiCallback<KeysBackupVersionTrust>(latch3) {
+                    override fun onSuccess(info: KeysBackupVersionTrust) {
+                        keysBackupVersionTrust = info
+                        super.onSuccess(info)
+                    }
+                })
+        mTestHelper.await(latch3)
+
+        // - It must be trusted and must have 2 signatures now
+        assertTrue(keysBackupVersionTrust!!.usable)
+        assertEquals(2, keysBackupVersionTrust!!.signatures.size)
+
+        stateObserver.stopAndCheckStates(null)
+        testData.cryptoTestData.clear(context)
+    }
+
+    /**
+     * - Do an e2e backup to the homeserver with a password
+     * - And log Alice on a new device
+     * - The new device must see the previous backup as not trusted
+     * - Try to trust the backup from the new device with a wrong password
+     * - It must fail
+     * - The backup must still be untrusted and disabled
+     */
+    @Test
+    fun trustKeyBackupVersionWithWrongPasswordTest() {
+        val password = "Password"
+        val badPassword = "Bad Password"
+
+        // - Do an e2e backup to the homeserver with a password
+        // - And log Alice on a new device
+        val context = InstrumentationRegistry.getContext()
+
+        val testData = createKeysBackupScenarioWithPassword(password, true)
+
+        val stateObserver = StateObserver(testData.aliceSession2.crypto!!.keysBackup)
+
+        // Wait for backup state to be NotTrusted
+        waitForKeysBackupToBeInState(testData.aliceSession2, KeysBackupStateManager.KeysBackupState.NotTrusted)
+
+        // - The new device must see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        // - Try to trust the backup from the new device with a wrong password
+        val latch = CountDownLatch(1)
+        testData.aliceSession2.crypto!!.keysBackup.trustKeysBackupVersionWithPassphrase(
+                testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion!!,
+                badPassword,
+                TestApiCallback(latch, false)
+        )
+        mTestHelper.await(latch)
+
+        // - The new device must still see the previous backup as not trusted
+        assertNotNull(testData.aliceSession2.crypto!!.keysBackup.mKeysBackupVersion)
+        assertFalse(testData.aliceSession2.crypto!!.keysBackup.isEnabled)
+        assertEquals(KeysBackupStateManager.KeysBackupState.NotTrusted, testData.aliceSession2.crypto!!.keysBackup.state)
+
+        stateObserver.stopAndCheckStates(null)
+        testData.cryptoTestData.clear(context)
+    }
+
     /**
      * - Do an e2e backup to the homeserver with a recovery key
      * - Log Alice on a new device
@@ -932,6 +1263,31 @@ class KeysBackupTest {
      * Private
      * ========================================================================================== */
 
+    /**
+     * As KaysBackup is doing asynchronous call to update its internal state, this method help to wait for the
+     * KeysBackup object to be in the specified state
+     */
+    private fun waitForKeysBackupToBeInState(mxSession: MXSession, state: KeysBackupStateManager.KeysBackupState) {
+        // If already in the wanted state, return
+        if (mxSession.crypto?.keysBackup?.state == state) {
+            return
+        }
+
+        // Else observe state changes
+        val latch = CountDownLatch(1)
+
+        mxSession.crypto?.keysBackup?.addListener(object : KeysBackupStateManager.KeysBackupStateListener {
+            override fun onStateChange(newState: KeysBackupStateManager.KeysBackupState) {
+                if (newState == state) {
+                    mxSession.crypto?.keysBackup?.removeListener(this)
+                    latch.countDown()
+                }
+            }
+        })
+
+        mTestHelper.await(latch)
+    }
+
     private data class PrepareKeysBackupDataResult(val megolmBackupCreationInfo: MegolmBackupCreationInfo,
                                                    val version: String)
 
@@ -1014,8 +1370,10 @@ class KeysBackupTest {
      * - Log Alice on a new device
      *
      * @param password optional password
+     * @param logoutFirstAliceSession set to true to logout first Alice session before opening the second one (default is false)
      */
-    private fun createKeysBackupScenarioWithPassword(password: String?): KeysBackupScenarioData {
+    private fun createKeysBackupScenarioWithPassword(password: String?,
+                                                     logoutFirstAliceSession: Boolean = false): KeysBackupScenarioData {
         val cryptoTestData = mCryptoTestHelper.doE2ETestWithAliceAndBobInARoomWithEncryptedMessages(true)
 
         val cryptoStore = cryptoTestData.firstSession.crypto!!.cryptoStore
@@ -1042,8 +1400,17 @@ class KeysBackupTest {
         assertEquals(2, lastProgress)
         assertEquals(2, lastTotal)
 
+        val aliceUserId = cryptoTestData.firstSession.myUserId
+
+        if (logoutFirstAliceSession) {
+            // Logout first Alice session, else they will share the same Crypto store and some tests may fail.
+            val latch2 = CountDownLatch(1)
+            cryptoTestData.firstSession.logout(InstrumentationRegistry.getContext(), TestApiCallback(latch2))
+            mTestHelper.await(latch2)
+        }
+
         // - Log Alice on a new device
-        val aliceSession2 = mTestHelper.logIntoAccount(cryptoTestData.firstSession.myUserId, defaultSessionParamsWithInitialSync)
+        val aliceSession2 = mTestHelper.logIntoAccount(aliceUserId, defaultSessionParamsWithInitialSync)
 
         // Test check: aliceSession2 has no keys at login
         assertEquals(0, aliceSession2.crypto!!.cryptoStore.inboundGroupSessionsCount(false))